moved Polygon class to regions.py added name as kwarg to testing regions functions added documentation for regions
