+++ csep/core/regions.py
@@ -5,40 +5,18 @@ from itertools import compress
 from xml.etree import ElementTree as ET
 
 # Third-party imports
+import matplotlib.path
 import numpy
 
 # PyCSEP imports
+import numpy as np
+import pyproj
+
 from csep.utils.calc import bin1d_vec
-from csep.utils.basic_types import Polygon
 from csep.utils.scaling_relationships import WellsAndCoppersmith
 
 
-def magnitude_bins(start_magnitude, end_magnitude, dmw):
-    """ Returns array holding magnitude bin edges.
-
-    The output from this function is monotonically increasing and equally spaced bin edges that can represent magnitude
-    bins.
-
-     Args:
-        start_magnitude (float)
-        end_magnitude (float)
-        dmw (float): magnitude spacing
-
-    Returns:
-        bin_edges (numpy.ndarray)
-    """
-    return numpy.arange(start_magnitude, end_magnitude+dmw/2, dmw)
-
-def create_space_magnitude_region(region, magnitudes):
-    """Simple wrapper to create space-magnitude region """
-    if not isinstance(region, CartesianGrid2D):
-        raise TypeError("region must be CartesianGrid2D")
-    # bind to region class
-    region.magnitudes = magnitudes
-    region.num_mag_bins = len(region.magnitudes)
-    return region
-
-def california_relm_region(dh_scale=1, magnitudes=None):
+def california_relm_region(dh_scale=1, magnitudes=None, name="relm-california"):
     """
     Returns class representing California testing region.
 
@@ -73,14 +51,14 @@ def california_relm_region(dh_scale=1, magnitudes=None):
 
     # turn points into polygons and make region object
     bboxes = compute_vertices(origins, dh)
-    relm_region = CartesianGrid2D([Polygon(bbox) for bbox in bboxes], dh, name="relm-california")
+    relm_region = CartesianGrid2D([Polygon(bbox) for bbox in bboxes], dh, name=name)
 
     if magnitudes is not None:
         relm_region.magnitudes = magnitudes
 
     return relm_region
 
-def italy_csep_region(dh_scale=1, magnitudes=None):
+def italy_csep_region(dh_scale=1, magnitudes=None, name="csep-italy"):
     """
         Returns class representing Italian testing region.
 
@@ -115,7 +93,7 @@ def italy_csep_region(dh_scale=1, magnitudes=None):
 
     # turn points into polygons and make region object
     bboxes = compute_vertices(origins, dh)
-    italy_region = CartesianGrid2D([Polygon(bbox) for bbox in bboxes], dh, name="csep-italy")
+    italy_region = CartesianGrid2D([Polygon(bbox) for bbox in bboxes], dh, name=name)
 
     if magnitudes is not None:
         italy_region.magnitudes = magnitudes
@@ -142,6 +120,31 @@ def global_region(dh=0.1, name="global", magnitudes=None):
         region.magnitudes = magnitudes
     return region
 
+def magnitude_bins(start_magnitude, end_magnitude, dmw):
+    """ Returns array holding magnitude bin edges.
+
+    The output from this function is monotonically increasing and equally spaced bin edges that can represent magnitude
+    bins.
+
+     Args:
+        start_magnitude (float)
+        end_magnitude (float)
+        dmw (float): magnitude spacing
+
+    Returns:
+        bin_edges (numpy.ndarray)
+    """
+    return numpy.arange(start_magnitude, end_magnitude+dmw/2, dmw)
+
+def create_space_magnitude_region(region, magnitudes):
+    """Simple wrapper to create space-magnitude region """
+    if not isinstance(region, CartesianGrid2D):
+        raise TypeError("region must be CartesianGrid2D")
+    # bind to region class
+    region.magnitudes = magnitudes
+    region.num_mag_bins = len(region.magnitudes)
+    return region
+
 def parse_csep_template(xml_filename):
     """
     Reads CSEP XML template file and returns the lat/lon values
@@ -166,7 +169,6 @@ def parse_csep_template(xml_filename):
         raise ValueError("dh_lat must equal dh_lon. grid needs to be regular.")
 
     return points, dh_lat
-
 
 def increase_grid_resolution(points, dh, factor):
     """
@@ -201,7 +203,6 @@ def increase_grid_resolution(points, dh, factor):
     new_factor = factor / 2
     return increase_grid_resolution(list(new_points), new_dh, new_factor)
 
-
 def masked_region(region, polygon):
     """
     Build a new region based off the coordinates in the polygon.
@@ -219,7 +220,6 @@ def masked_region(region, polygon):
     new_polygons = list(compress(region.polygons, contains))
     # create new region with the spatial cells inside the polygon
     return CartesianGrid2D(new_polygons, region.dh)
-
 
 def generate_aftershock_region(mainshock_mw, mainshock_lon, mainshock_lat, num_radii=3, region=california_relm_region, **kwargs):
     """ Creates a spatial region around a given epicenter
@@ -244,6 +244,264 @@ def generate_aftershock_region(mainshock_mw, mainshock_lon, mainshock_lat, num_r
                                                           num_radii * rupture_length, num_points=100)
     aftershock_region = masked_region(region(**kwargs), aftershock_polygon)
     return aftershock_region
+
+def grid_spacing(vertices):
+    """
+    Figures out the length and
+
+    Args:
+        vertices: Vertices describe a single node in grid.
+
+    Returns:
+        dh: grid spacing
+
+    Raises:
+        ValueError
+
+    """
+    # get first two vertices
+    a = vertices[0]
+    b = vertices[1]
+    # compute both differences, because unless point is the same one is bound to be the dh
+    d1 = numpy.abs(b[0] - a[0])
+    d2 = numpy.abs(b[1] - a[1])
+    if not numpy.allclose(d1, d2):
+        raise ValueError("grid spacing must be regular for cartesian grid.")
+    dh = numpy.max([d1, d2])
+    # this would happen if the same point is repeated twice
+    if dh == 0:
+        raise ValueError("Problem computing grid spacing cannot be zero.")
+    return dh
+
+def compute_vertex(origin_point, dh, tol=numpy.finfo(float).eps):
+    """
+    Computes the bounding box of a rectangular polygon given its origin points and spacing dh.
+
+    Args:
+        origin_points: list of tuples, where tuple is (x, y)
+        dh: spacing
+        tol: used to eliminate overlapping polygons in the case of a rectangular mesh, defaults to
+             the machine tolerance.
+
+    Returns:
+        list of polygon edges
+
+    """
+    bbox = ((origin_point[0], origin_point[1]),
+            (origin_point[0], origin_point[1] + dh - tol),
+            (origin_point[0] + dh - tol, origin_point[1] + dh - tol),
+            (origin_point[0] + dh - tol, origin_point[1]))
+    return bbox
+
+def compute_vertices(origin_points, dh, tol=numpy.finfo(float).eps):
+    """
+    Wrapper function to compute vertices for multiple points. Default tolerance is set to machine precision
+    of floating point number.
+
+    Args:
+        origin_points: 2d ndarray
+
+    Notes:
+        (x,y) should be accessible like:
+        >>> x_coords = origin_points[:,0]
+        >>> y_coords = origin_points[:,1]
+
+    """
+    return list(map(lambda x: compute_vertex(x, dh, tol=tol), origin_points))
+
+def _build_bitmask_vec(polygons, dh):
+    """
+    same as build mask but using vectorized calls to bin1d
+    """
+    # build bounding box of set of polygons based on origins
+    nd_origins = numpy.array([poly.origin for poly in polygons])
+    bbox = [(numpy.min(nd_origins[:, 0]), numpy.min(nd_origins[:, 1])),
+            (numpy.max(nd_origins[:, 0]), numpy.max(nd_origins[:, 1]))]
+
+    # get midpoints for hashing
+    midpoints = numpy.array([poly.centroid() for poly in polygons])
+
+    # compute nx and ny
+    nx = numpy.rint((bbox[1][0] - bbox[0][0]) / dh)
+    ny = numpy.rint((bbox[1][1] - bbox[0][1]) / dh)
+
+    # set up grid of bounding box
+    xs = dh * numpy.arange(nx + 1) + bbox[0][0]
+    ys = dh * numpy.arange(ny + 1) + bbox[0][1]
+
+    # set up mask array, 1 is index 0 is mask
+    a = numpy.ones([len(ys), len(xs), 2])
+
+    # set all indices to nan
+    a[:,:,1] = numpy.nan
+
+    # bin1d returns the index of polygon within the cartesian grid
+    idx = bin1d_vec(midpoints[:, 0], xs)
+    idy = bin1d_vec(midpoints[:, 1], ys)
+
+    for i in range(len(polygons)):
+        # store index of polygon in dim=1
+        a[idy[i], idx[i], 1] = int(i)
+        # build mask in dim=0; here we masked values are 1
+        if idx[i] >= 0 and idy[i] >= 0:
+            a[idy[i], idx[i], 0] = 0
+
+    return a, xs, ys
+
+def _bin_catalog_spatio_magnitude_counts(lons, lats, mags, n_poly, mask, idx_map, binx, biny, mag_bins):
+    """
+    Returns a list of event counts as ndarray with shape (n_poly, n_cat) where each value
+    represents the event counts within the polygon.
+
+    Using [:, :, 1] index of the mask, we store the mapping between the index of n_poly and
+    that polygon in the mask. Additionally, the polygons are ordered such that the index of n_poly
+    in the result corresponds to the index of the polygons.
+
+    Eventually, we can make a structure that could contain both of these, but the trade-offs will need
+    to be compared against performance.
+    """
+
+    # index in cartesian grid for events in data. note, this has a different index than the
+    # vector of polygons. this mapping is stored in [:,:,1] index of mask
+    # index in 2d grid
+    idx = bin1d_vec(lons, binx)
+    idy = bin1d_vec(lats, biny)
+    mag_idxs = bin1d_vec(mags, mag_bins, right_continuous=True)
+    # start with zero event counts in each bin
+    event_counts = numpy.zeros((n_poly, len(mag_bins)))
+    # does not seem that we can vectorize this part
+    skipped = []
+    for i in range(idx.shape[0]):
+        if not mask[idy[i], idx[i]] and idy[i] != -1 and idx[i] != -1 and mag_idxs[i] != -1:
+            # getting spatial bin from mask
+            hash_idx = int(idx_map[idy[i], idx[i]])
+            mag_idx = mag_idxs[i]
+            # update event counts in that polygon
+            event_counts[(hash_idx, mag_idx)] += 1
+        else:
+            skipped.append((lons[i], lats[i], mags[i]))
+
+    return event_counts, skipped
+
+def _bin_catalog_spatial_counts(lons, lats, n_poly, mask, idx_map, binx, biny):
+    """
+    Returns a list of event counts as ndarray with shape (n_poly) where each value
+    represents the event counts within the polygon.
+
+    Using [:, :, 1] index of the mask, we store the mapping between the index of n_poly and
+    that polygon in the mask. Additionally, the polygons are ordered such that the index of n_poly
+    in the result corresponds to the index of the polygons.
+
+    We can make a structure that could contain both of these, but the trade-offs will need
+    to be compared against performance.
+    """
+    ai, bi = binx, biny
+    # will return negative
+    idx = bin1d_vec(lons, ai)
+    idy = bin1d_vec(lats, bi)
+    # bin1d returns -1 if outside the region
+    # todo: think about how to change this behavior for less confusions, bc -1 is an actual value that can be chosen
+    bad = (idx == -1) | (idy == -1) | (mask[idy,idx] == 1)
+    # this can be memory optimized by keeping short list and storing index, only for case where n/2 events
+    event_counts = numpy.zeros(n_poly)
+    # selecting the indexes into polygons correspoding to lons and lats within the grid
+    hash_idx = idx_map[idy[~bad],idx[~bad]].astype(int)
+    # aggregate in counts
+    numpy.add.at(event_counts, hash_idx, 1)
+    return event_counts
+
+def _bin_catalog_probability(lons, lats, n_poly, mask, idx_map, binx, biny):
+    """
+    Returns a list of event counts as ndarray with shape (n_poly) where each value
+    represents the event counts within the polygon.
+
+    Using [:, :, 1] index of the mask, we store the mapping between the index of n_poly and
+    that polygon in the mask. Additionally, the polygons are ordered such that the index of n_poly
+    in the result corresponds to the index of the polygons.
+
+    We can make a structure that could contain both of these, but the trade-offs will need
+    to be compared against performance.
+    """
+    ai, bi = binx, biny
+    # returns -1 if outside of the bbox
+    idx = bin1d_vec(lons, ai)
+    idy = bin1d_vec(lats, bi)
+    bad = (idx == -1) | (idy == -1) | (mask[idy, idx] == 1)
+    event_counts = numpy.zeros(n_poly)
+    # [:,:,1] is a mapping from the polygon array to cartesian grid
+    hash_idx = idx_map[idy[~bad],idx[~bad]].astype(int)
+    # dont accumulate just set to one for probability
+    event_counts[hash_idx] = 1
+    return event_counts
+
+class Polygon:
+    """
+    Represents polygons defined through a collection of vertices.
+
+    This polygon is assumed to be 2d, but could contain an arbitrary number of vertices. The path is treated as not being
+    closed.
+    """
+    def __init__(self, points):
+        # instance members
+        self.points = points
+        self.origin = self.points[0]
+
+        # https://matplotlib.org/3.1.1/api/path_api.html
+        self.path = matplotlib.path.Path(self.points)
+
+    def __str__(self):
+        return str(self.origin)
+
+    def contains(self, points):
+        """ Returns a bool array which is True if the path contains the corresponding point.
+
+        Args:
+            points: 2d numpy array
+
+        """
+        nd_points = np.array(points)
+        if nd_points.ndim == 1:
+            nd_points = nd_points.reshape(1,-1)
+        return self.path.contains_points(nd_points)
+
+    def centroid(self):
+        """ return the centroid of the polygon."""
+        c0, c1 = 0, 0
+        k = len(self.points)
+        for p in self.points:
+            c0 = c0 + p[0]
+            c1 = c1 + p[1]
+        return c0 / k, c1 / k
+
+    def get_xcoords(self):
+        return np.array(self.points)[:,0]
+
+    def get_ycoords(self):
+        return np.array(self.points)[:,1]
+
+    @classmethod
+    def from_great_circle_radius(cls, centroid, radius, num_points=10):
+        """
+        Generates a polygon object from a given radius and centroid location.
+
+        Args:
+            centroid: (lon, lat)
+            radius: should be in (meters)
+            num_points: more points is higher resolution polygon
+
+        Returns:
+            polygon
+        """
+        geod = pyproj.Geod(ellps='WGS84')
+        azim = np.linspace(0, 360, num_points)
+        # create vectors with same length as azim for computations
+        center_lons = np.ones(num_points) * centroid[0]
+        center_lats = np.ones(num_points) * centroid[1]
+        radius = np.ones(num_points) * radius
+        # get new lons and lats
+        endlon, endlat, backaz = geod.fwd(center_lons, center_lats, azim, radius)
+        # class method
+        return cls(np.column_stack([endlon, endlat]))
 
 
 class CartesianGrid2D:
@@ -271,11 +529,11 @@ class CartesianGrid2D:
 
     @property
     def num_nodes(self):
+        """ Number of polygons in region """
         return len(self.polygons)
 
     def get_index_of(self, lons, lats):
-        """
-        Returns the index of lons, lats in self.polygons
+        """ Returns the index of lons, lats in self.polygons
 
         Args:
             lons: ndarray-like
@@ -324,7 +582,10 @@ class CartesianGrid2D:
 
         idx = bin1d_vec(lons, self.xs)
         idy = bin1d_vec(lats, self.ys)
-        # todo: needs error checking for -1 values
+        if numpy.any(idx == -1) or numpy.any(idy == -1):
+            raise ValueError("at least one lon and lat pair contain values that are outside of the valid region.")
+        if numpy.any(self.mask[idy, idx] == 1):
+            raise ValueError("at least one lon and lat pair contain values that are outside of the valid region.")
         return self.mask[idy, idx].astype(bool)
 
     def get_cartesian(self, data):
@@ -347,12 +608,15 @@ class CartesianGrid2D:
         return results
 
     def get_bbox(self):
+        """ Returns rectangular bounding box around region. """
         return (self.xs.min(), self.xs.max(), self.ys.min(), self.ys.max())
 
     def midpoints(self):
+        """ Returns midpoints of rectangular polygons in region """
         return numpy.array([poly.centroid() for poly in self.polygons])
 
     def origins(self):
+        """ Returns origins of rectangular polygons in region """
         return numpy.array([poly.origin for poly in self.polygons])
 
     def to_dict(self):
@@ -368,7 +632,7 @@ class CartesianGrid2D:
         raise NotImplementedError("Todo!")
 
     @classmethod
-    def from_origins(cls, origins, magnitudes=None, name=None):
+    def from_origins(cls, origins, dh=None, magnitudes=None, name=None):
         """Creates instance of class from 2d numpy.array of lon/lat origins.
 
         Note: Grid spacing should be constant in the entire region. This condition is not explicitly checked for for performance
@@ -389,207 +653,13 @@ class CartesianGrid2D:
             raise TypeError("origins must be of type numpy.array or be numpy array like.")
 
         # dh must be regular, no explicit checking.
-        dh2 = numpy.abs(lons[1]-lons[0])
-        dh1 = numpy.abs(lats[1]-lats[0])
-        dh = numpy.max([dh1, dh2])
+        if dh is None:
+            dh2 = numpy.abs(lons[1]-lons[0])
+            dh1 = numpy.abs(lats[1]-lats[0])
+            dh = numpy.max([dh1, dh2])
 
         region = CartesianGrid2D([Polygon(bbox) for bbox in compute_vertices(origins, dh)], dh, name=name)
         if magnitudes is not None:
             region.magnitudes = magnitudes
         return region
 
-
-def grid_spacing(vertices):
-    """
-    Figures out the length and
-
-    Args:
-        vertices: Vertices describe a single node in grid.
-
-    Returns:
-        dh: grid spacing
-
-    Raises:
-        ValueError
-
-    """
-    # get first two vertices
-    a = vertices[0]
-    b = vertices[1]
-    # compute both differences, because unless point is the same one is bound to be the dh
-    d1 = numpy.abs(b[0] - a[0])
-    d2 = numpy.abs(b[1] - a[1])
-    if not numpy.allclose(d1, d2):
-        raise ValueError("grid spacing must be regular for cartesian grid.")
-    dh = numpy.max([d1, d2])
-    # this would happen if the same point is repeated twice
-    if dh == 0:
-        raise ValueError("Problem computing grid spacing cannot be zero.")
-    return dh
-
-
-def compute_vertex(origin_point, dh, tol=numpy.finfo(float).eps):
-    """
-    Computes the bounding box of a rectangular polygon given its origin points and spacing dh.
-
-    Args:
-        origin_points: list of tuples, where tuple is (x, y)
-        dh: spacing
-        tol: used to eliminate overlapping polygons in the case of a rectangular mesh, defaults to
-             the machine tolerance.
-
-    Returns:
-        list of polygon edges
-
-    """
-    bbox = ((origin_point[0], origin_point[1]),
-            (origin_point[0], origin_point[1] + dh - tol),
-            (origin_point[0] + dh - tol, origin_point[1] + dh - tol),
-            (origin_point[0] + dh - tol, origin_point[1]))
-    return bbox
-
-
-def compute_vertices(origin_points, dh, tol=numpy.finfo(float).eps):
-    """
-    Wrapper function to compute vertices for multiple points. Default tolerance is set to machine precision
-    of floating point number.
-
-    Args:
-        origin_points: 2d ndarray
-
-    Notes:
-        (x,y) should be accessible like:
-        >>> x_coords = origin_points[:,0]
-        >>> y_coords = origin_points[:,1]
-
-    """
-    return list(map(lambda x: compute_vertex(x, dh, tol=tol), origin_points))
-
-
-def _build_bitmask_vec(polygons, dh):
-    """
-    same as build mask but using vectorized calls to bin1d
-    """
-    # build bounding box of set of polygons based on origins
-    nd_origins = numpy.array([poly.origin for poly in polygons])
-    bbox = [(numpy.min(nd_origins[:, 0]), numpy.min(nd_origins[:, 1])),
-            (numpy.max(nd_origins[:, 0]), numpy.max(nd_origins[:, 1]))]
-
-    # get midpoints for hashing
-    midpoints = numpy.array([poly.centroid() for poly in polygons])
-
-    # compute nx and ny
-    nx = numpy.rint((bbox[1][0] - bbox[0][0]) / dh)
-    ny = numpy.rint((bbox[1][1] - bbox[0][1]) / dh)
-
-    # set up grid of bounding box
-    xs = dh * numpy.arange(nx + 1) + bbox[0][0]
-    ys = dh * numpy.arange(ny + 1) + bbox[0][1]
-
-    # set up mask array, 1 is index 0 is mask
-    a = numpy.ones([len(ys), len(xs), 2])
-
-    # set all indices to nan
-    a[:,:,1] = numpy.nan
-
-    # bin1d returns the index of polygon within the cartesian grid
-    idx = bin1d_vec(midpoints[:, 0], xs)
-    idy = bin1d_vec(midpoints[:, 1], ys)
-
-    for i in range(len(polygons)):
-        # store index of polygon in dim=1
-        a[idy[i], idx[i], 1] = int(i)
-        # build mask in dim=0; here we masked values are 1
-        if idx[i] >= 0 and idy[i] >= 0:
-            a[idy[i], idx[i], 0] = 0
-
-    return a, xs, ys
-
-
-def _bin_catalog_spatio_magnitude_counts(lons, lats, mags, n_poly, mask, idx_map, binx, biny, mag_bins):
-    """
-    Returns a list of event counts as ndarray with shape (n_poly, n_cat) where each value
-    represents the event counts within the polygon.
-
-    Using [:, :, 1] index of the mask, we store the mapping between the index of n_poly and
-    that polygon in the mask. Additionally, the polygons are ordered such that the index of n_poly
-    in the result corresponds to the index of the polygons.
-
-    Eventually, we can make a structure that could contain both of these, but the trade-offs will need
-    to be compared against performance.
-    """
-
-    # index in cartesian grid for events in data. note, this has a different index than the
-    # vector of polygons. this mapping is stored in [:,:,1] index of mask
-    # index in 2d grid
-    idx = bin1d_vec(lons, binx)
-    idy = bin1d_vec(lats, biny)
-    mag_idxs = bin1d_vec(mags, mag_bins, right_continuous=True)
-    # start with zero event counts in each bin
-    event_counts = numpy.zeros((n_poly, len(mag_bins)))
-    # does not seem that we can vectorize this part
-    skipped = []
-    for i in range(idx.shape[0]):
-        if not mask[idy[i], idx[i]] and idy[i] != -1 and idx[i] != -1 and mag_idxs[i] != -1:
-            # getting spatial bin from mask
-            hash_idx = int(idx_map[idy[i], idx[i]])
-            mag_idx = mag_idxs[i]
-            # update event counts in that polygon
-            event_counts[(hash_idx, mag_idx)] += 1
-        else:
-            skipped.append((lons[i], lats[i], mags[i]))
-
-    return event_counts, skipped
-
-
-def _bin_catalog_spatial_counts(lons, lats, n_poly, mask, idx_map, binx, biny):
-    """
-    Returns a list of event counts as ndarray with shape (n_poly) where each value
-    represents the event counts within the polygon.
-
-    Using [:, :, 1] index of the mask, we store the mapping between the index of n_poly and
-    that polygon in the mask. Additionally, the polygons are ordered such that the index of n_poly
-    in the result corresponds to the index of the polygons.
-
-    We can make a structure that could contain both of these, but the trade-offs will need
-    to be compared against performance.
-    """
-    ai, bi = binx, biny
-    # will return negative
-    idx = bin1d_vec(lons, ai)
-    idy = bin1d_vec(lats, bi)
-    # bin1d returns -1 if outside the region
-    # todo: think about how to change this behavior for less confusions, bc -1 is an actual value that can be chosen
-    bad = (idx == -1) | (idy == -1) | (mask[idy,idx] == 1)
-    # this can be memory optimized by keeping short list and storing index, only for case where n/2 events
-    event_counts = numpy.zeros(n_poly)
-    # selecting the indexes into polygons correspoding to lons and lats within the grid
-    hash_idx = idx_map[idy[~bad],idx[~bad]].astype(int)
-    # aggregate in counts
-    numpy.add.at(event_counts, hash_idx, 1)
-    return event_counts
-
-
-def _bin_catalog_probability(lons, lats, n_poly, mask, idx_map, binx, biny):
-    """
-    Returns a list of event counts as ndarray with shape (n_poly) where each value
-    represents the event counts within the polygon.
-
-    Using [:, :, 1] index of the mask, we store the mapping between the index of n_poly and
-    that polygon in the mask. Additionally, the polygons are ordered such that the index of n_poly
-    in the result corresponds to the index of the polygons.
-
-    We can make a structure that could contain both of these, but the trade-offs will need
-    to be compared against performance.
-    """
-    ai, bi = binx, biny
-    # returns -1 if outside of the bbox
-    idx = bin1d_vec(lons, ai)
-    idy = bin1d_vec(lats, bi)
-    bad = (idx == -1) | (idy == -1) | (mask[idy, idx] == 1)
-    event_counts = numpy.zeros(n_poly)
-    # [:,:,1] is a mapping from the polygon array to cartesian grid
-    hash_idx = idx_map[idy[~bad],idx[~bad]].astype(int)
-    # dont accumulate just set to one for probability
-    event_counts[hash_idx] = 1
-    return event_counts
